requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
greenlet==1.1.2
idna==3.2
msal==1.14.0
numpy==1.21.2
Office365-REST-Python-Client==2.3.8
pandas==1.3.3
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
